"Consolidate event listeners"
FINALLY figured it out!

I killed the 3 event listeners
and replaced them with one on
the parent element.

I forgot that event listeners pass
the event data to the function (how?!).

This helped fix the order in which events
occured, making my initial program
work for the most part.

I